refactor: typeorm category entity 파일 이름 변경
