+++ src/modules/posts/entities/post.entity.ts
@@ -2,7 +2,7 @@ import 'reflect-metadata';
 import { Column, Entity, JoinColumn, ManyToOne, OneToMany } from 'typeorm';
 import { User } from '../../users/entities/user.entity';
 import { BaseModel } from '../../../config/typeorm/base.entity';
-import { Category } from '../../categorys/entities/category.entity';
+import { Category } from '../../categorys/entities/category-typeorm.entity';
 import { Comment } from '../../comments/entities/comment.entity';
 import { PostImage } from './post-images.entity';
